scrapper-alexis/requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
tabulate>=0.9.0
lxml>=4.9.0
requests>=2.31.0

